[Fix]지출 리스트 페이지 데이터 연동 수정

+++ WanderBoard/pond1225/ExpenseModel.swift
@@ -11,7 +11,7 @@ import Foundation
 struct Expense {
     let date : Date
     let expenseContent : String
-    let expenseAmount : Double
+    let expenseAmount : Int
     let category : String
     let memo : String
     let imageName: String
